final changes and cleanup

+++ src/main/java/com/github/dmitriims/posikengine/service/DatabaseService.java
@@ -11,6 +11,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import javax.persistence.Tuple;
+import java.math.BigInteger;
 import java.time.LocalDateTime;
 import java.util.*;
 import java.util.stream.Collectors;
@@ -39,23 +40,24 @@ public class DatabaseService {
     }
 
     @Transactional
-    public void saveOrUpdatePage(Site site, Page page, List<Lemma> originalLemmas, CommonContext commonContext) {
-        Page pageToReindex = pageRepository.findBySiteAndPathEquals(site, page.getPath());
+    public void saveOrUpdatePage(Page page, List<Lemma> originalLemmas, CommonContext commonContext) {
+        Page pageToReindex = pageRepository.findBySiteIdAndPagePath(page.getSite().getId(), page.getPath());
         if (pageToReindex == null) {
-            savePageToDataBase(site, page, originalLemmas, commonContext);
+            savePageToDataBase(page, originalLemmas, commonContext);
             return;
         }
         if (page.equals(pageToReindex)) {
             addPageToSavedPagesMap(pageToReindex);
+            setSiteStatusToIndexing(page.getSite().getId());
             return;
         }
         dropOldIndexesAndDecrementLemmasFrequencies(pageToReindex.getId());
-        savePageToDataBase(site, page, originalLemmas, commonContext);
+        savePageToDataBase(page, originalLemmas, commonContext);
     }
 
     @Transactional
-    public void savePageToDataBase(Site site, Page page, List<Lemma> lemmas, CommonContext commonContext) {
-        Page savedPage = pageRepository.findBySiteAndPathEquals(site, page.getPath());
+    public void savePageToDataBase(Page page, List<Lemma> lemmas, CommonContext commonContext) {
+        Page savedPage = pageRepository.findBySiteIdAndPagePath(page.getSite().getId(), page.getPath());
         if (savedPage == null) {
             savedPage = pageRepository.save(page);
         } else if (!savedPage.equals(page)) {
@@ -67,7 +69,7 @@ public class DatabaseService {
         }
         addPageToSavedPagesMap(savedPage);
         saveNewLemmasAndIndexes(savedPage, lemmas);
-        setIndexingStatusOnCompletion(site, commonContext);
+        setIndexingStatusOnCompletion(page.getSite().getId(), commonContext);
     }
 
     private void addPageToSavedPagesMap(Page page) {
@@ -78,20 +80,25 @@ public class DatabaseService {
     }
 
     @Transactional
-    public void removeDeletedPagesForSite(Long siteId) {
+    public boolean removeDeletedPagesForSite(Long siteId) {
         Set<Long> savedPagesSet = savedPagesPerSite.get(siteId);
+        if(savedPagesSet == null) {
+            savedPagesPerSite.remove(siteId);
+            return false;
+        }
         Set<Long> pagesInDb = new HashSet<>(pageRepository.getAllIdsBySiteId(Collections.singletonList(siteId)));
         if (pagesInDb.size() > savedPagesSet.size()) {
-            log.info("cleaning up deleted pages for site " + siteRepository.findById(siteId));
+            log.info("удаляю из базы данных страницы, которых больше нет на сайте " + siteRepository.findById(siteId).orElseThrow().getUrl());
             pagesInDb.removeAll(savedPagesSet);
             for (Long pageId : pagesInDb) {
                 dropOldIndexesAndDecrementLemmasFrequencies(pageId);
                 pageRepository.deleteById(pageId);
             }
-            log.info("cleanup complete");
+            log.info("удаление лишних страниц завершено");
 
         }
         savedPagesPerSite.remove(siteId);
+        return true;
     }
 
     @Transactional
@@ -117,7 +124,7 @@ public class DatabaseService {
     }
 
     private void saveNewLemmasAndIndexes(Page page, List<Lemma> newLemmas) {
-        List<Lemma> lemmasFromDB = lemmaRepository.findAllBySiteAndLemmaIn(page.getSite(), newLemmas.stream().map(Lemma::getLemma).collect(Collectors.toList()));
+        List<Lemma> lemmasFromDB = getLemmasFromDBByLemmasAndSiteId(page.getSite().getId(), newLemmas.stream().map(Lemma::getLemma).collect(Collectors.toList()));
         List<Lemma> savedLemmas = saveLemmas(newLemmas, lemmasFromDB);
         saveIndexes(page, newLemmas, savedLemmas);
     }
@@ -158,57 +165,41 @@ public class DatabaseService {
     }
 
     @Transactional
-    public Site setSiteStatusToIndexed(Site site) {
-        Site siteToUpdate = siteRepository.findById(site.getId()).get();
+    public void setSiteStatusToIndexed(Long siteId) {
+        Site siteToUpdate = siteRepository.findById(siteId).get();
         siteToUpdate.setStatus(Status.INDEXED);
         siteToUpdate.setLastError("");
         siteToUpdate.setStatusTime(LocalDateTime.now());
-        site = siteRepository.saveAndFlush(siteToUpdate);
-        return site;
+        siteRepository.saveAndFlush(siteToUpdate);
     }
 
     @Transactional
-    public Site setSiteStatusToIndexing(Site site) {
-        Site siteToUpdate = siteRepository.findById(site.getId()).get();
+    public void setSiteStatusToIndexing(Long siteId) {
+        Site siteToUpdate = siteRepository.findById(siteId).get();
         siteToUpdate.setStatus(Status.INDEXING);
         siteToUpdate.setLastError("");
         siteToUpdate.setStatusTime(LocalDateTime.now());
-        site = siteRepository.saveAndFlush(siteToUpdate);
-        return site;
+        siteRepository.saveAndFlush(siteToUpdate);
     }
 
     @Transactional
-    public void setSiteStatusToFailed(Site site, String error) {
-        Site siteToUpdate = siteRepository.findById(site.getId()).get();
+    public void setSiteStatusToFailed(Long siteId, String error) {
+        Site siteToUpdate = siteRepository.findById(siteId).get();
         siteToUpdate.setStatus(Status.FAILED);
         siteToUpdate.setLastError(error);
         siteToUpdate.setStatusTime(LocalDateTime.now());
         siteRepository.saveAndFlush(siteToUpdate);
     }
 
-    @Transactional
-    public void setAllSiteStatusesToFailed(String error) {
-        List<Site> sites = siteRepository.findAll();
-        sites.forEach(s -> setSiteStatusToFailed(s, error));
-    }
 
-    private void setIndexingStatusOnCompletion(Site site, CommonContext commonContext) {
+    private void setIndexingStatusOnCompletion(Long siteId, CommonContext commonContext) {
         if (commonContext.isIndexing()) {
-            setSiteStatusToIndexing(site);
+            setSiteStatusToIndexing(siteId);
         } else {
-            setSiteStatusToFailed(site, commonContext.getIndexingMessage());
+            setSiteStatusToFailed(siteId, commonContext.getIndexingMessage());
         }
     }
 
-    @Transactional
-    public void deleteSiteInformation(Site site) {
-        List<Long> pageIds = pageRepository.getAllIdsBySiteId(Collections.singletonList(site.getId()));
-        for (long id : pageIds) {
-            indexRepository.deleteAllByPage_Id(id);
-        }
-        lemmaRepository.deleteAllBySite(site);
-        pageRepository.deleteAllBySite(site);
-    }
 
     @Transactional
     public List<String> filterPopularLemmasOut(List<Site> sites, List<String> lemmas, double threshold) {
@@ -241,6 +232,17 @@ public class DatabaseService {
                         t.get(2, String.class),
                         t.get(3, String.class),
                         t.get(4, Double.class)))
+                .collect(Collectors.toList());
+    }
+
+    private List<Lemma> getLemmasFromDBByLemmasAndSiteId(Long siteId, List<String> lemmas) {
+        return lemmaRepository.findAllBySiteAndLemmaIn(siteId, lemmas).stream()
+                .map(t -> new Lemma(
+                        t.get(0, BigInteger.class).longValue(),
+                        siteRepository.findById(t.get(3, BigInteger.class).longValue()).get(),
+                        t.get(2, String.class),
+                        t.get(1, Integer.class)
+                ))
                 .collect(Collectors.toList());
     }
